pequenas alterações

+++ churchmanager/src/main/java/br/com/churchmanager/model/group/PessoaAtividaEclesiastica.java
@@ -1,14 +1,15 @@
 package br.com.churchmanager.model.group;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public class PessoaAtividaEclesiastica implements Serializable {
 	private static final long serialVersionUID = 3026082101675912254L;
 	private String categoria;
-	private BigInteger quantidade;
+	private BigDecimal quantidade;
 
-	public PessoaAtividaEclesiastica(String categoria, BigInteger quantidade) {
+	public PessoaAtividaEclesiastica(String categoria, BigDecimal quantidade) {
 		this.categoria = categoria;
 		this.quantidade = quantidade;
 	}
@@ -21,11 +22,11 @@ public class PessoaAtividaEclesiastica implements Serializable {
 		this.categoria = categoria;
 	}
 
-	public BigInteger getQuantidade() {
+	public BigDecimal getQuantidade() {
 		return this.quantidade;
 	}
 
-	public void setQuantidade(BigInteger quantidade) {
+	public void setQuantidade(BigDecimal quantidade) {
 		this.quantidade = quantidade;
 	}
 }
